Тарасова ДЮ, ПРОГ - 5, ЛР -5/web_client/requirements.txt
flask==2.3.3
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2